use f32 (we're using float 32 arrays anyway)

+++ src/kinematic.rs
@@ -3,7 +3,7 @@ use crate::vector::Vector;
 #[derive(Clone, Copy)]
 pub struct Kinematic {
     pub position: Vector,
-    pub orientation: f64,
+    pub orientation: f32,
     pub velocity: Vector,
-    pub rotation: f64,
+    pub rotation: f32,
 }
